+++ src/AbstractFilepond.php
@@ -1,7 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace RahulHaque\Filepond;
 
+use const UPLOAD_ERR_OK;
+
+use Exception;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Crypt;
@@ -21,6 +26,24 @@ abstract class AbstractFilepond
     private $isOwnershipAware;
 
     private $isSoftDeletable;
+
+    /**
+     * @return string
+     */
+    public function getTempDisk()
+    {
+        return $this->tempDisk;
+    }
+
+    /**
+     * @return $this
+     */
+    public function setTempDisk(string $tempDisk)
+    {
+        $this->tempDisk = $tempDisk;
+
+        return $this;
+    }
 
     /**
      * Decrypt the FilePond field value data
@@ -62,24 +85,6 @@ abstract class AbstractFilepond
         }
 
         $this->fieldValue = $this->decrypt($fieldValue);
-
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getTempDisk()
-    {
-        return $this->tempDisk;
-    }
-
-    /**
-     * @return $this
-     */
-    public function setTempDisk(string $tempDisk)
-    {
-        $this->tempDisk = $tempDisk;
 
         return $this;
     }
@@ -196,11 +201,17 @@ abstract class AbstractFilepond
      */
     protected function createFileObject(Filepond $filepond)
     {
+        $driver = config('filesystems.disks.'.$this->tempDisk.'.driver');
+
+        if ($driver !== 'local') {
+            throw new Exception('Unable to create file object for ['.$this->tempDisk.'] disk driver.');
+        }
+
         return new UploadedFile(
             Storage::disk($this->tempDisk)->path($filepond->filepath),
             $filepond->filename,
             $filepond->mimetypes,
-            \UPLOAD_ERR_OK,
+            UPLOAD_ERR_OK,
             true
         );
     }