+++ src/FilepondServiceProvider.php
@@ -1,10 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
 namespace RahulHaque\Filepond;
 
+use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Support\ServiceProvider;
 use Illuminate\Validation\Rule;
 use RahulHaque\Filepond\Console\FilepondClear;
+use RahulHaque\Filepond\Factories\UploaderManager;
 use RahulHaque\Filepond\Rules\FilepondRule;
 
 class FilepondServiceProvider extends ServiceProvider
@@ -25,11 +29,9 @@ class FilepondServiceProvider extends ServiceProvider
                 __DIR__.'/../config/filepond.php' => base_path('config/filepond.php'),
             ], 'filepond-config');
 
-            if (! class_exists('CreateFilepondsTable')) {
-                $this->publishes([
-                    __DIR__.'/../database/migrations/create_fileponds_table.php.stub' => database_path('migrations/'.date('Y_m_d', time()).'_000000_create_fileponds_table.php'),
-                ], 'filepond-migrations');
-            }
+            $this->publishes([
+                __DIR__.'/../database/migrations/create_fileponds_table.php.stub' => glob(database_path('/migrations/*_create_fileponds_table.php'))[0] ?? database_path('migrations/'.date('Y_m_d_His', time()).'_create_fileponds_table.php'),
+            ], 'filepond-migrations');
 
             $this->commands([
                 FilepondClear::class,
@@ -44,8 +46,8 @@ class FilepondServiceProvider extends ServiceProvider
     {
         $this->mergeConfigFrom(__DIR__.'/../config/filepond.php', 'filepond');
 
-        $this->app->singleton('filepond', function () {
-            return new Filepond;
-        });
+        $this->app->singleton(UploaderManager::class, fn (Application $app) => new UploaderManager($app));
+
+        $this->app->singleton('filepond', fn () => new Filepond);
     }
 }