+++ src/Rules/FilepondRule.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace RahulHaque\Filepond\Rules;
 
 use Illuminate\Contracts\Validation\DataAwareRule;