+++ tests/Feature/FilepondFacadeTest.php
@@ -1,20 +1,20 @@
 <?php
 
+declare(strict_types=1);
+
 namespace RahulHaque\Filepond\Tests\Feature;
 
-use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Storage;
+use PHPUnit\Framework\Attributes\Test;
 use RahulHaque\Filepond\Facades\Filepond;
 use RahulHaque\Filepond\Tests\TestCase;
 use RahulHaque\Filepond\Tests\User;
 
 class FilepondFacadeTest extends TestCase
 {
-    use RefreshDatabase;
-
-    /** @test */
-    public function can_get_temporary_file_after_filepond_file_upload()
+    #[Test]
+    public function can_get_temporary_file_after_file_upload()
     {
         Storage::disk(config('filepond.temp_disk', 'local'))->deleteDirectory(config('filepond.temp_folder', 'filepond/temp'));
 
@@ -28,7 +28,7 @@ class FilepondFacadeTest extends TestCase
                 'avatar' => $uploadedFile,
             ], [
                 'Content-Type' => 'multipart/form-data',
-                'accept' => 'application/json',
+                'Accept' => 'application/json',
             ]);
 
         $temporaryFile = Filepond::field($response->content())->getFile();
@@ -36,8 +36,8 @@ class FilepondFacadeTest extends TestCase
         $this->assertEquals($temporaryFile->getSize(), $uploadedFile->getSize());
     }
 
-    /** @test */
-    public function can_get_data_url_after_filepond_file_upload()
+    #[Test]
+    public function can_get_data_url_after_file_upload()
     {
         Storage::disk(config('filepond.temp_disk', 'local'))->deleteDirectory(config('filepond.temp_folder', 'filepond/temp'));
 
@@ -52,7 +52,7 @@ class FilepondFacadeTest extends TestCase
                 'avatar' => $uploadedFile,
             ], [
                 'Content-Type' => 'multipart/form-data',
-                'accept' => 'application/json',
+                'Accept' => 'application/json',
             ]);
 
         $dataUrl = Filepond::field($response->content())->getDataURL();
@@ -60,8 +60,8 @@ class FilepondFacadeTest extends TestCase
         $this->assertEquals($uploadEncoded, last(explode(',', $dataUrl)));
     }
 
-    /** @test */
-    public function can_copy_filepond_file_upload_to_desired_location()
+    #[Test]
+    public function can_copy_file_upload_to_desired_location()
     {
         Storage::disk(config('filepond.temp_disk', 'local'))->deleteDirectory(config('filepond.temp_folder', 'filepond/temp'));
 
@@ -73,7 +73,7 @@ class FilepondFacadeTest extends TestCase
                 'avatar' => UploadedFile::fake()->image('avatar.png', 1024, 1024),
             ], [
                 'Content-Type' => 'multipart/form-data',
-                'accept' => 'application/json',
+                'Accept' => 'application/json',
             ]);
 
         $fileInfo = Filepond::field($response->content())->copyTo('avatars/avatar-1');
@@ -81,8 +81,8 @@ class FilepondFacadeTest extends TestCase
         Storage::disk(config('filepond.disk', 'local'))->assertExists($fileInfo['location']);
     }
 
-    /** @test */
-    public function can_move_filepond_file_upload_to_desired_location()
+    #[Test]
+    public function can_move_file_upload_to_desired_location()
     {
         Storage::disk(config('filepond.temp_disk', 'local'))->deleteDirectory(config('filepond.temp_folder', 'filepond/temp'));
 
@@ -94,7 +94,7 @@ class FilepondFacadeTest extends TestCase
                 'avatar' => UploadedFile::fake()->image('avatar.png', 1024, 1024),
             ], [
                 'Content-Type' => 'multipart/form-data',
-                'accept' => 'application/json',
+                'Accept' => 'application/json',
             ]);
 
         $fileInfo = Filepond::field($response->content())->moveTo('avatars/avatar-1');
@@ -102,8 +102,8 @@ class FilepondFacadeTest extends TestCase
         Storage::disk(config('filepond.disk', 'local'))->assertExists($fileInfo['location']);
     }
 
-    /** @test */
-    public function can_copy_multiple_filepond_file_upload_to_desired_location()
+    #[Test]
+    public function can_copy_multiple_file_upload_to_desired_location()
     {
         Storage::disk(config('filepond.temp_disk', 'local'))->deleteDirectory(config('filepond.temp_folder', 'filepond/temp'));
 
@@ -118,7 +118,7 @@ class FilepondFacadeTest extends TestCase
                     'gallery' => UploadedFile::fake()->image('gallery-'.$i.'.png', 1024, 1024),
                 ], [
                     'Content-Type' => 'multipart/form-data',
-                    'accept' => 'application/json',
+                    'Accept' => 'application/json',
                 ]);
 
             $request[] = $response->content();
@@ -131,8 +131,8 @@ class FilepondFacadeTest extends TestCase
         }
     }
 
-    /** @test */
-    public function can_move_multiple_filepond_file_upload_to_desired_location()
+    #[Test]
+    public function can_move_multiple_file_upload_to_desired_location()
     {
         Storage::disk(config('filepond.temp_disk', 'local'))->deleteDirectory(config('filepond.temp_folder', 'filepond/temp'));
 
@@ -147,7 +147,7 @@ class FilepondFacadeTest extends TestCase
                     'gallery' => UploadedFile::fake()->image('gallery-'.$i.'.png', 1024, 1024),
                 ], [
                     'Content-Type' => 'multipart/form-data',
-                    'accept' => 'application/json',
+                    'Accept' => 'application/json',
                 ]);
 
             $request[] = $response->content();