+++ tests/Unit/FilepondProcessRouteTest.php
@@ -1,20 +1,20 @@
 <?php
 
+declare(strict_types=1);
+
 namespace RahulHaque\Filepond\Tests\Unit;
 
-use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Crypt;
 use Illuminate\Support\Facades\Storage;
+use PHPUnit\Framework\Attributes\Test;
 use RahulHaque\Filepond\Models\Filepond;
 use RahulHaque\Filepond\Tests\TestCase;
 use RahulHaque\Filepond\Tests\User;
 
 class FilepondProcessRouteTest extends TestCase
 {
-    use RefreshDatabase;
-
-    /** @test */
+    #[Test]
     public function can_validate_filepond_file_upload_request()
     {
         Storage::disk(config('filepond.temp_disk', 'local'))->deleteDirectory(config('filepond.temp_folder', 'filepond/temp'));
@@ -27,13 +27,13 @@ class FilepondProcessRouteTest extends TestCase
                 'avatar' => 'string_input_instead_of_file',
             ], [
                 'Content-Type' => 'multipart/form-data',
-                'accept' => 'application/json',
+                'Accept' => 'application/json',
             ]);
 
         $response->assertJson(['avatar' => ['The avatar field must be a file.']]);
     }
 
-    /** @test */
+    #[Test]
     public function can_process_filepond_file_upload_request()
     {
         Storage::disk(config('filepond.temp_disk', 'local'))->deleteDirectory(config('filepond.temp_folder', 'filepond/temp'));
@@ -46,7 +46,7 @@ class FilepondProcessRouteTest extends TestCase
                 'avatar' => UploadedFile::fake()->image('avatar.png', 100, 100),
             ], [
                 'Content-Type' => 'multipart/form-data',
-                'accept' => 'application/json',
+                'Accept' => 'application/json',
             ]);
 
         $data = Crypt::decrypt($response->content());
@@ -56,7 +56,7 @@ class FilepondProcessRouteTest extends TestCase
         Storage::disk(config('filepond.temp_disk', 'local'))->assertExists($fileById->filepath);
     }
 
-    /** @test */
+    #[Test]
     public function can_process_filepond_array_file_upload_request()
     {
         Storage::disk(config('filepond.temp_disk', 'local'))->deleteDirectory(config('filepond.temp_folder', 'filepond/temp'));
@@ -69,7 +69,7 @@ class FilepondProcessRouteTest extends TestCase
                 'gallery' => ['profile' => UploadedFile::fake()->image('avatar.png', 1024, 1024)],
             ], [
                 'Content-Type' => 'multipart/form-data',
-                'accept' => 'application/json',
+                'Accept' => 'application/json',
             ]);
 
         $data = Crypt::decrypt($response->content());